+++ R/forest.netcomb.R
@@ -333,7 +333,7 @@ forest.netcomb <- function(x,
     suppressWarnings(metagen(TE, seTE, data = dat,
                              sm = x$sm,
                              studlab = trts, backtransf = backtransf,
-                             method.tau = "DL",
+                             method.tau = "DL", method.tau.ci = "",
                              warn = FALSE))
   ##
   forest(m1,