+++ R/funnel.netmeta.R
@@ -428,7 +428,7 @@ funnel.netmeta <- function(x,
   m.adj <-
     suppressWarnings(metagen(res$TE.adj, res$seTE,
                              studlab = res$studlab, sm = x$sm,
-                             method.tau = "DL"))
+                             method.tau = "DL", method.tau.ci = ""))
   ##
   n.comps <- length(unique(res$comparison))
   ##