+++ R/pairwise.R
@@ -957,7 +957,8 @@ pairwise <- function(treat,
                           dat$event2, dat$n2,
                           incr = dat$incr, addincr = TRUE,
                           allstudies = allstudies,
-                          warn = warn, method.tau.ci = "", ...)
+                          method.tau = "DL", method.tau.ci = "",
+                          warn = warn, ...)
             ##
             dat$TE   <- m1$TE
             dat$seTE <- m1$seTE
@@ -1074,9 +1075,9 @@ pairwise <- function(treat,
           if (nrow(dat) > 0) {
             m1 <- metacont(dat$n1, dat$mean1, dat$sd1,
                            dat$n2, dat$mean2, dat$sd2,
-                           warn = warn,
-                           method.tau.ci = "",
-                           method.smd = "Cohen", ...)
+                           method.tau = "DL", method.tau.ci = "",
+                           method.smd = "Cohen",
+                           warn = warn, ...)
             ##
             dat$TE   <- m1$TE
             dat$seTE <- m1$seTE
@@ -1286,7 +1287,8 @@ pairwise <- function(treat,
                           dat$event2, dat$time2,
                           incr = dat$incr, addincr = TRUE,
                           allstudies = allstudies,
-                          warn = warn, method.tau.ci = "", ...)
+                          method.tau = "DL", method.tau.ci = "",
+                          warn = warn, ...)
             ##
             dat$TE <- m1$TE
             dat$seTE <- m1$seTE
@@ -1433,18 +1435,38 @@ pairwise <- function(treat,
   
   
   ##
-  ## Use first treatment as reference if argument is missing
+  ## Use first treatment with estimable effect as reference if
+  ## argument is missing
   ##
   labels <- unique(sort(c(res$treat1, res$treat2)))
   ##
-  if (missing.reference.group)
-    reference.group <- labels[1]
+  if (missing.reference.group) {
+    go.on <- TRUE
+    i <- 0
+    while (go.on) {
+      i <- i + 1
+      sel.i <-
+        !is.na(res$TE) & !is.na(res$seTE) &
+        (res$treat1 == labels[i] | res$treat2 == labels[i])
+      if (sum(sel.i) > 0) {
+        go.on <- FALSE
+        reference.group <- labels[i]
+      }
+      else if (i == length(labels)) {
+        go.on <- FALSE
+        reference.group <- ""
+      }
+    }
+  }
+  ##
   if (is.factor(reference.group))
     reference.group <- as.character(reference.group)
+  ##
   if (is.numeric(reference.group))
     chknumeric(reference.group, length = 1)
   else
     chkchar(reference.group, length = 1)
+  ##
   reference.group <- setchar(reference.group, c(labels, ""))
   ##
   if (!keep.all.comparisons) {
@@ -1456,7 +1478,7 @@ pairwise <- function(treat,
       trts.i <- unique(sort(c(d.i$treat1, d.i$treat2)))
       ##
       ## Keep comparisons with reference group or first treatment if
-      ## reference treament missing in study
+      ## reference treatment is missing in study
       ##
       if (reference.group %in% trts.i)
         ref.i <- reference.group